feat(faz-21-6-prc1): migrate chart imports to x-charts in consumers
Apply the PR-C1 codemod to runtime widgets and update non-runtime
metadata so the canonical chart import is `@mfe/x-charts`.

Codemod-driven migration (3 widgets):
- apps/mfe-shell/.../ContextHealthTrendWidget.tsx — LineChart
- apps/mfe-shell/.../WorkIntakeWidget.tsx — PieChart
- apps/mfe-shell/.../ExtensionHealthWidget.tsx — BarChart

Each widget had its DS chart import split into two declarations:
non-chart specifiers stay on `@mfe/design-system`, the chart
specifier moves to `@mfe/x-charts`.

Manual non-runtime metadata updates:
- packages/design-system-docs/.../quick-start.mdx — code-block
  import sample updated to x-charts
- apps/mfe-shell/.../design-lab.items.part-01.v1.json —
  importStatement strings for AreaChart/BarChart catalogue entries
- apps/mfe-shell/.../design-lab.items.part-04.v1.json —
  importStatement strings for LineChart/PieChart catalogue entries

Driveby fix to the codemod itself (output now matches the repo's
single-quote prettier style):
- packages/design-system/scripts/codemods/
  migrate-ds-charts-to-x-charts.ts —
  `root.toSource({ quote: 'single' })` (was 'double')

Acceptance:
- `rg "@mfe/design-system" packages/x-charts/src` -> 0 hits
- `rg ".../widgets/{Context,Work,Ext}.tsx | grep Chart"` -> 0 hits
- `rg "import \\w+ from .../@mfe/design-system | grep Chart"` -> 0
- pnpm --filter @mfe/x-charts test -> 1008 passed (57 files)
- pnpm --filter @mfe/design-system test -> 8450 passed (532 files)
- pnpm --filter mfe-shell exec tsc --noEmit -> exit 0

Refs: Faz 21.6 PR-C1 LITE (Codex iter-3 AGREE)

+++ packages/design-system/scripts/codemods/migrate-ds-charts-to-x-charts.ts
@@ -72,7 +72,7 @@ const transform = (file: FileInfo, api: API, _options: Options): string | null =
       mutated = true;
     });
 
-  return mutated ? root.toSource({ quote: 'double' }) : null;
+  return mutated ? root.toSource({ quote: 'single' }) : null;
 };
 
 export default transform;
